+++ libmsip/source/headers/SipHeader.cxx
@@ -179,8 +179,8 @@ MRef<SipHeader *> SipHeader::parseHeader(const string &line){
 				value_zero = value_params[0];
 // 				cerr<<"valueline.substr(2): "+valueline.substr(2)<<endl;
 			} else if( headerType == "Contact" ) {
-				uint32_t ltPos = (uint32_t) values[i].find( '<' );
-				uint32_t gtPos = (uint32_t) values[i].find( '>' );
+				size_t ltPos = values[i].find( '<' );
+				size_t gtPos = values[i].find( '>' );
 				if( ltPos!=string::npos && gtPos!=string::npos ) {
 					//if the string contains '<' and '>' ...
 					//remove them
@@ -198,8 +198,8 @@ MRef<SipHeader *> SipHeader::parseHeader(const string &line){
 				}
 			} else if( headerType == "From" ||
 				   headerType == "To" ) {
-				uint32_t ltPos = (uint32_t) values[i].find( '<' );
-				uint32_t gtPos = (uint32_t) values[i].find( '>' );
+				size_t ltPos = values[i].find( '<' );
+				size_t gtPos = values[i].find( '>' );
 				if( ltPos!=string::npos && gtPos!=string::npos ) {
 					value_zero = values[i].substr( 0, gtPos + 1 );
 					//now split the outside parameters ...
